pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibm-watsonx-orchestrate"
version = "0.1.0"
description = "IBM watsonx.orchestrate SDK"
authors = [
    { name = "IBM", email = "support@ibm.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2024.8.30",
    "importlib-metadata>=8.5.0",
    "packaging>=24.2",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
    "urllib3>=2.2.3",
    "pydantic>=2.10.3,<3.0.0",
    "langgraph>=0.2.58,<0.3.0",
    "pyyaml>=6.0.2,<7.0.0",
    "docstring-parser>=0.16,<1.0",
    "httpx>=0.28.1,<1.0.0",
    "jsonschema>=4.23.0,<5.0.0",
    "loguru>=0.7.3,<1.0.0",
    "langchain-community>=0.3.12,<1.0.0",
    "rich>=13.9.4,<14.0.0",
    "pyjwt>=2.10.1,<3.0.0",
    "typer>=0.15.1,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "snapshottest==1.0.0a1",
    "pytest-mock==3.14.0",
    "pytest-asyncio==0.25.1",
    "coverage[toml]>=6.5",
]

[tool.hatch.envs.default]
dependencies = [
    "pkg[dev]"
]
detached = true

[tool.hatch.build.targets.wheel]
packages = ["src/ibm_watsonx_orchestrate"]

[project.scripts]
orchestrate = "ibm_watsonx_orchestrate.cli.main:app"

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[tool.coverage.run]
omit = [
  "*/__init__.py",
  "tests/*"
]

[tool.coverage.report]
show_missing=true
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
